requirements.txt
fastapi
uvicorn
python-dotenv

# LangChain Modular Packages
langchain==0.2.1
langchain-core
langchain-community
langchain-openai
langchain-astradb

# Vector DB & Embeddings
chromadb  # only needed if you want to test Chroma fallback
openai

# Optional Dev Tools
tqdm

